+++ esp-hal/src/soc/mod.rs
@@ -6,6 +6,7 @@ use portable_atomic::AtomicU32;
 use procmacros::ram;
 
 pub use self::implementation::*;
+use crate::efuse::ChipRevision;
 
 #[cfg_attr(esp32, path = "esp32/mod.rs")]
 #[cfg_attr(esp32c2, path = "esp32c2/mod.rs")]
@@ -332,26 +333,29 @@ pub(crate) fn setup_trap_section_protection() {
 
 static CHIP_REVISION: AtomicU32 = AtomicU32::new(0);
 const LOADED: u32 = 1 << 31;
+const MAX_REVISION: ChipRevision = ChipRevision::from_packed(0xFFFF);
 
 #[cold]
 fn load_chip_revision_from_efuse() -> u16 {
     let chip_revision = crate::efuse::chip_revision();
+    let chip_revision = chip_revision.packed();
     CHIP_REVISION.store(chip_revision as u32 | LOADED, Ordering::Release);
     chip_revision
 }
 
 #[ram]
-fn load_chip_revision() -> u16 {
+fn load_chip_revision() -> ChipRevision {
     let stored = CHIP_REVISION.load(Ordering::Acquire);
     if stored & LOADED == 0 {
-        return load_chip_revision_from_efuse();
+        return ChipRevision::from_packed(load_chip_revision_from_efuse());
     }
-    (stored & u16::MAX as u32) as u16
+    ChipRevision::from_packed((stored & u16::MAX as u32) as u16)
 }
 
-fn chip_revision_in_range(range: Range<u16>) -> bool {
-    const BUILD_TIME_MIN_REV: u16 =
-        esp_config::esp_config_int!(u16, "ESP_HAL_CONFIG_MIN_CHIP_REVISION");
+fn chip_revision_in_range(range: Range<ChipRevision>) -> bool {
+    const BUILD_TIME_MIN_REV: ChipRevision = ChipRevision::from_combined(
+        esp_config::esp_config_int!(u16, "ESP_HAL_CONFIG_MIN_CHIP_REVISION"),
+    );
 
     // Check to determine chip is obviously in or out of range, without reading efuse
     #[allow(
@@ -367,24 +371,27 @@ fn chip_revision_in_range(range: Range<u16>) -> bool {
         clippy::absurd_extreme_comparisons,
         reason = "Not absurd depending on configuration"
     )]
-    if range.start <= BUILD_TIME_MIN_REV && range.end == u16::MAX {
+    if range.start <= BUILD_TIME_MIN_REV && range.end == MAX_REVISION {
         return true;
     }
 
     let chip_revision = load_chip_revision();
 
-    range.contains(&chip_revision)
+    range.start <= chip_revision && chip_revision < range.end
 }
 
 /// Returns true if the chip revision is at least the given revision.
 #[allow(dead_code)]
-pub(crate) fn chip_revision_above(min: u16) -> bool {
-    chip_revision_in_range(min..u16::MAX)
+pub(crate) fn chip_revision_above(revision: ChipRevision) -> bool {
+    chip_revision_in_range(revision..MAX_REVISION)
 }
 
 /// Returns true if the chip is at least the given revision, in the same major version.
 #[allow(dead_code)]
-pub(crate) fn chip_minor_revision_above(rev: u16) -> bool {
-    let max = (rev + 1).next_multiple_of(100);
-    chip_revision_in_range(rev..max)
+pub(crate) fn chip_minor_revision_above(revision: ChipRevision) -> bool {
+    let next_major = ChipRevision {
+        major: revision.major + 1,
+        minor: 0,
+    };
+    chip_revision_in_range(revision..next_major)
 }