+++ esp-hal/src/uart/mod.rs
@@ -69,7 +69,7 @@ use crate::{
     handler,
     interrupt::InterruptHandler,
     pac::uart0::RegisterBlock,
-    private::OnDrop,
+    private::DropGuard,
     ram,
     soc::clocks::{self, ClockTree},
     system::PeripheralGuard,
@@ -162,11 +162,8 @@ impl embedded_io_07::Error for TxError {
     }
 }
 
-#[cfg(feature = "unstable")]
-#[cfg_attr(docsrs, doc(cfg(feature = "unstable")))]
+#[instability::unstable]
 pub use crate::soc::clocks::UartFunctionClockSclk as ClockSource;
-#[cfg(not(feature = "unstable"))]
-use crate::soc::clocks::UartFunctionClockSclk as ClockSource;
 use crate::soc::clocks::{
     UartBaudRateGeneratorConfig as BaudRateConfig,
     UartFunctionClockConfig as ClockConfig,
@@ -1087,7 +1084,7 @@ impl<'d> UartRx<'d, Async> {
             // future resolved.
             let info = self.uart.info();
             unwrap!(info.set_rx_fifo_full_threshold(max_threshold));
-            let _guard = OnDrop::new(|| {
+            let _guard = DropGuard::new((), |_| {
                 unwrap!(info.set_rx_fifo_full_threshold(current_threshold));
             });