+++ esp-hal/CHANGELOG.md
@@ -53,6 +53,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ESP32-C61: RNG (#5244)
 - C61: Add GPIO support (#5248)
 - C61: Add UART support (#5251)
+- C61: Add I2C support (#5258)
+- C61: Add SPI support (#5261)
+- Added support for `rand_core 0.10.0` (#5280)
+- `esp_hal::efuse::ChipRevision` (#5287)
 
 ### Changed
 
@@ -89,6 +93,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - S3: SPI1 is no longer initialized if PSRAM is not correctly detected. The warning message now includes PSRAM mode config (#5122)
 - `Rng`, `Rng::{random, read}` have been marked stable (#5098)
 - `esp_hal::init` now verifies that the initial stack pointer is in range (#5227)
+- Updated embassy dependencies: embassy-sync to 0.8, embassy-embedded-hal to 0.6 (#5249)
+- `esp_hal::efuse::chip_revision` has been marked stable (#5287)
+- `esp_hal::efuse::chip_revision` now returns `ChipRevision` (#5287)
 
 ### Fixed